Move pairplot and jointplot
