+++ seaborn/distributions.py
@@ -15,12 +15,11 @@ try:
 except ImportError:
     _has_statsmodels = False
 
-from .utils import set_hls_values, iqr, _kde_support
+from .utils import iqr, _kde_support
 from .palettes import color_palette, blend_palette
-from .axisgrid import JointGrid
 
 
-__all__ = ["distplot", "kdeplot", "rugplot", "jointplot"]
+__all__ = ["distplot", "kdeplot", "rugplot"]
 
 
 def _freedman_diaconis_bins(a):
@@ -668,228 +667,3 @@ def rugplot(a, height=.05, axis="x", ax=None, **kwargs):
         func(pt, 0, height, **kwargs)
 
     return ax
-
-
-def jointplot(x, y, data=None, kind="scatter", stat_func=stats.pearsonr,
-              color=None, size=6, ratio=5, space=.2,
-              dropna=True, xlim=None, ylim=None,
-              joint_kws=None, marginal_kws=None, annot_kws=None, **kwargs):
-    """Draw a plot of two variables with bivariate and univariate graphs.
-
-    This function provides a convenient interface to the :class:`JointGrid`
-    class, with several canned plot kinds. This is intended to be a fairly
-    lightweight wrapper; if you need more flexibility, you should use
-    :class:`JointGrid` directly.
-
-    Parameters
-    ----------
-    x, y : strings or vectors
-        Data or names of variables in ``data``.
-    data : DataFrame, optional
-        DataFrame when ``x`` and ``y`` are variable names.
-    kind : { "scatter" | "reg" | "resid" | "kde" | "hex" }, optional
-        Kind of plot to draw.
-    stat_func : callable or None, optional
-        Function used to calculate a statistic about the relationship and
-        annotate the plot. Should map `x` and `y` either to a single value
-        or to a (value, p) tuple. Set to ``None`` if you don't want to
-        annotate the plot.
-    color : matplotlib color, optional
-        Color used for the plot elements.
-    size : numeric, optional
-        Size of the figure (it will be square).
-    ratio : numeric, optional
-        Ratio of joint axes size to marginal axes height.
-    space : numeric, optional
-        Space between the joint and marginal axes
-    dropna : bool, optional
-        If True, remove observations that are missing from ``x`` and ``y``.
-    {x, y}lim : two-tuples, optional
-        Axis limits to set before plotting.
-    {joint, marginal, annot}_kws : dicts, optional
-        Additional keyword arguments for the plot components.
-    kwargs : key, value pairings
-        Additional keyword arguments are passed to the function used to
-        draw the plot on the joint Axes, superseding items in the
-        ``joint_kws`` dictionary.
-
-    Returns
-    -------
-    grid : :class:`JointGrid`
-        :class:`JointGrid` object with the plot on it.
-
-    See Also
-    --------
-    JointGrid : The Grid class used for drawing this plot. Use it directly if
-                you need more flexibility.
-
-    Examples
-    --------
-
-    Draw a scatterplot with marginal histograms:
-
-    .. plot::
-        :context: close-figs
-
-        >>> import numpy as np, pandas as pd; np.random.seed(0)
-        >>> import seaborn as sns; sns.set(style="white", color_codes=True)
-        >>> tips = sns.load_dataset("tips")
-        >>> g = sns.jointplot(x="total_bill", y="tip", data=tips)
-
-    Add regression and kernel density fits:
-
-    .. plot::
-        :context: close-figs
-
-        >>> g = sns.jointplot("total_bill", "tip", data=tips, kind="reg")
-
-    Replace the scatterplot with a joint histogram using hexagonal bins:
-
-    .. plot::
-        :context: close-figs
-
-        >>> g = sns.jointplot("total_bill", "tip", data=tips, kind="hex")
-
-    Replace the scatterplots and histograms with density estimates and align
-    the marginal Axes tightly with the joint Axes:
-
-    .. plot::
-        :context: close-figs
-
-        >>> iris = sns.load_dataset("iris")
-        >>> g = sns.jointplot("sepal_width", "petal_length", data=iris,
-        ...                   kind="kde", space=0, color="g")
-
-    Use a different statistic for the annotation:
-
-    .. plot::
-        :context: close-figs
-
-        >>> from scipy.stats import spearmanr
-        >>> g = sns.jointplot("size", "total_bill", data=tips,
-        ...                   stat_func=spearmanr, color="m")
-
-    Draw a scatterplot, then add a joint density estimate:
-
-    .. plot::
-        :context: close-figs
-
-        >>> g = (sns.jointplot("sepal_length", "sepal_width",
-        ...                    data=iris, color="k")
-        ...         .plot_joint(sns.kdeplot, zorder=0, n_levels=6))
-
-    Pass vectors in directly without using Pandas, then name the axes:
-
-    .. plot::
-        :context: close-figs
-
-        >>> x, y = np.random.randn(2, 300)
-        >>> g = (sns.jointplot(x, y, kind="hex", stat_func=None)
-        ...         .set_axis_labels("x", "y"))
-
-    Draw a smaller figure with more space devoted to the marginal plots:
-
-    .. plot::
-        :context: close-figs
-
-        >>> g = sns.jointplot("total_bill", "tip", data=tips,
-        ...                   size=5, ratio=3, color="g")
-
-    Pass keyword arguments down to the underlying plots:
-
-    .. plot::
-        :context: close-figs
-
-        >>> g = sns.jointplot("petal_length", "sepal_length", data=iris,
-        ...                   marginal_kws=dict(bins=15, rug=True),
-        ...                   annot_kws=dict(stat="r"),
-        ...                   s=40, edgecolor="w", linewidth=1)
-
-    """
-    # Set up empty default kwarg dicts
-    if joint_kws is None:
-        joint_kws = {}
-    joint_kws.update(kwargs)
-    if marginal_kws is None:
-        marginal_kws = {}
-    if annot_kws is None:
-        annot_kws = {}
-
-    # Make a colormap based off the plot color
-    if color is None:
-        color = color_palette()[0]
-    color_rgb = mpl.colors.colorConverter.to_rgb(color)
-    colors = [set_hls_values(color_rgb, l=l) for l in np.linspace(1, 0, 12)]
-    cmap = blend_palette(colors, as_cmap=True)
-
-    # Initialize the JointGrid object
-    grid = JointGrid(x, y, data, dropna=dropna,
-                     size=size, ratio=ratio, space=space,
-                     xlim=xlim, ylim=ylim)
-
-    # Plot the data using the grid
-    if kind == "scatter":
-
-        joint_kws.setdefault("color", color)
-        grid.plot_joint(plt.scatter, **joint_kws)
-
-        marginal_kws.setdefault("kde", False)
-        marginal_kws.setdefault("color", color)
-        grid.plot_marginals(distplot, **marginal_kws)
-
-    elif kind.startswith("hex"):
-
-        x_bins = min(_freedman_diaconis_bins(grid.x), 50)
-        y_bins = min(_freedman_diaconis_bins(grid.y), 50)
-        gridsize = int(np.mean([x_bins, y_bins]))
-
-        joint_kws.setdefault("gridsize", gridsize)
-        joint_kws.setdefault("cmap", cmap)
-        grid.plot_joint(plt.hexbin, **joint_kws)
-
-        marginal_kws.setdefault("kde", False)
-        marginal_kws.setdefault("color", color)
-        grid.plot_marginals(distplot, **marginal_kws)
-
-    elif kind.startswith("kde"):
-
-        joint_kws.setdefault("shade", True)
-        joint_kws.setdefault("cmap", cmap)
-        grid.plot_joint(kdeplot, **joint_kws)
-
-        marginal_kws.setdefault("shade", True)
-        marginal_kws.setdefault("color", color)
-        grid.plot_marginals(kdeplot, **marginal_kws)
-
-    elif kind.startswith("reg"):
-
-        from .regression import regplot
-
-        marginal_kws.setdefault("color", color)
-        grid.plot_marginals(distplot, **marginal_kws)
-
-        joint_kws.setdefault("color", color)
-        grid.plot_joint(regplot, **joint_kws)
-
-    elif kind.startswith("resid"):
-
-        from .regression import residplot
-
-        joint_kws.setdefault("color", color)
-        grid.plot_joint(residplot, **joint_kws)
-
-        x, y = grid.ax_joint.collections[0].get_offsets().T
-        marginal_kws.setdefault("color", color)
-        marginal_kws.setdefault("kde", False)
-        distplot(x, ax=grid.ax_marg_x, **marginal_kws)
-        distplot(y, vertical=True, fit=stats.norm, ax=grid.ax_marg_y,
-                 **marginal_kws)
-        stat_func = None
-    else:
-        msg = "kind must be either 'scatter', 'reg', 'resid', 'kde', or 'hex'"
-        raise ValueError(msg)
-
-    if stat_func is not None:
-        grid.annotate(stat_func, **annot_kws)
-
-    return grid
